@@ -1,7 +1,7 @@
 name: audioplayers
-description: A flutter plugin to play multiple audio files simultaneously
-version: 0.15.1
-homepage: https://github.com/luanpotter/audioplayer
+description: A Flutter plugin to play multiple audio files simultaneously
+version: 0.18.1
+homepage: https://github.com/luanpotter/audioplayers
 
 flutter:
   plugin:
@@ -18,8 +18,8 @@ flutter:
         fileName: audioplayers_web.dart
 
 dependencies:
-  uuid: ^2.0.4
-  path_provider: ^1.5.1
+  uuid: ^3.0.1
+  path_provider: ^2.0.1
   flutter:
     sdk: flutter
   flutter_web_plugins:
@@ -28,10 +28,9 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  http: ^0.12.0+3
-  test: ^1.9.4
+  http: ^0.13.0
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
-  flutter: ">=1.10.0 <2.0.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=1.10.0"
 
