@@ -1,4 +1,1 @@
-#import <FlutterMacOS/FlutterMacOS.h>
-
-@interface AudioplayersPlugin : NSObject<FlutterPlugin>
-@end
+../../darwin/Classes/AudioplayersPlugin.h
